+++ packages/opencode/src/agent/agent.ts
@@ -18,7 +18,7 @@ import { mergeDeep, pipe, sortBy, values } from "remeda"
 import { Global } from "@/global"
 import path from "path"
 import { Plugin } from "@/plugin"
-import { Skill } from "../skill"
+import { Skill } from "../skill/skill"
 import { Effect, Context, Layer } from "effect"
 import { InstanceState } from "@/effect"
 import * as Option from "effect/Option"
