+++ packages/opencode/src/effect/app-runtime.ts
@@ -18,7 +18,7 @@ import { Plugin } from "@/plugin"
 import { Provider } from "@/provider"
 import { ProviderAuth } from "@/provider"
 import { Agent } from "@/agent/agent"
-import { Skill } from "@/skill"
+import { Skill } from "@/skill/skill"
 import { Discovery } from "@/skill/discovery"
 import { Question } from "@/question"
 import { Permission } from "@/permission"
