+++ packages/opencode/src/server/instance/index.ts
@@ -8,7 +8,7 @@ import { TuiRoutes } from "./tui"
 import { Instance } from "../../project/instance"
 import { Vcs } from "../../project"
 import { Agent } from "../../agent/agent"
-import { Skill } from "../../skill"
+import { Skill } from "../../skill/skill"
 import { Global } from "../../global"
 import { LSP } from "../../lsp"
 import { Command } from "../../command"
