--- packages/opencode/src/skill/index.ts
@@ -1,1 +1,0 @@
-export * as Skill from "./skill"
