+++ packages/opencode/src/skill/skill.ts
@@ -260,3 +260,5 @@ export function fmt(list: Info[], opts: { verbose: boolean }) {
       .map((skill) => `- **${skill.name}**: ${skill.description}`),
   ].join("\n")
 }
+
+export * as Skill from "./skill"
