+++ packages/opencode/test/session/snapshot-tool-race.test.ts
@@ -41,7 +41,7 @@ import { Plugin } from "../../src/plugin"
 import { Provider as ProviderSvc } from "../../src/provider"
 import { Env } from "../../src/env"
 import { Question } from "../../src/question"
-import { Skill } from "../../src/skill"
+import { Skill } from "../../src/skill/skill"
 import { SystemPrompt } from "../../src/session/system"
 import { Todo } from "../../src/session/todo"
 import { SessionCompaction } from "../../src/session/compaction"
